started injection paging library feature for competition short list, added dependencies for paging library, created needed classes

+++ app/src/main/java/com/rodionov/osport/data/dto/requests/CompetitionRequest.kt
@@ -17,7 +17,7 @@ data class CompetitionRequest(
     @SerializedName("place")
     val place: Place,
     @SerializedName("start_interval")
-    val startInterval: Float,
+    val startInterval: Double,
     @SerializedName("organizers")
     val organizers: List<OrganizerDto>,
     @SerializedName("title")
